test(cypress): consolidate lifecycle coverage
